+++ test/com/jec/juta/annotations/AfterClassFailTest.ts
@@ -14,15 +14,16 @@
 //   See the License for the specific language governing permissions and
 //   limitations under the License.
 
-/*!
- * This module constains utilities used by the AnnotatedMethodTypeTest test
- * suite.
- */
+import "mocha";
+import {expect} from "chai";
 
 // Utilities:
-export const BEFORE_CLASS:number = 0;
-export const AFTER_CLASS:number = 1;
-export const BEFORE_ALL:number = 2;
-export const AFTER_ALL:number = 3;
-export const BEFORE:number = 4;
-export const AFTER:number = 5;
+import * as utils from "../../../../../utils/test-utils/utilities/AfterClassFailTestUtils";
+
+// Test:
+describe("@AfterClass", ()=> {
+
+  it("should throw an error since no context has been declared for this decorator", function() {
+    expect(utils.buildClassRef).to.throw(Error);
+  });
+});