+++ utils/test-utils/classes/annotations/BeforeClassTestClass.ts
@@ -1,0 +1,8 @@
+import {BeforeClass} from "../../../../src/com/jec/juta/annotations/BeforeClass";
+import * as params from "../../utilities/ParamUtils";
+
+export class BeforeClassTestClass {
+
+  @BeforeClass(params.ANNOTATED_METHOD_PARAMS)
+  public static beforeClassMethod():void { }
+}